updated name conventions
